pyproject.toml
[tool.poetry]
name = "kivymd_reload_async"
version = "0.1.3"
description = "Demo KivyMD Python GUI app with hot reload and trio async"
authors = ["Ernest Prabhakar <ernest@quiltdata.io>"]
license = "MIT"
readme = "README.md"
packages = [{include = "kivymd_reload_async"}]

[tool.poetry.dependencies]
python = "^3.10"
kivymd = "^1.1.1"
trio = "^0.22.0"
httpx = "^0.23.3"

[tool.poetry.group.dev.dependencies]
watchdog = "^2.2.1"
pytest-watch = "^4.2.0"
flake8 = "^6.0.0"
pytest-trio = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
